src/pictionary-app/requirements.txt
boto3==1.24.87
numpy==1.26.4
pandas==2.1.0
flask==3.0.2
pillow==10.2.0
torch==2.2.1
torchvision==0.17.1